Added ability to represent multiple groups of the same tag (see #2, #3)
This change adds an ability to represent messages with repeated
groups of attributes with the same group tag. The most noticeable
use case is the Get-Jobs response which uses multiple Job groups,
one per returned job. See RFC 8011, 4.2.6.2. for more details.

With respect to backward compatibility, the following
behavior is implemented here:
  1. (*Message).Decode() fills both Groups and named per-group
     fields (i.e., Operation, Job etc)
  2. (*Message).Encode() and (*Message) Print, if Groups != nil,
     uses Groups and ignores  named per-group fields. Otherwise,
     named fields are used as in 1.0.0
  3. (*Message) Equal(), for each message uses Groups if
     it is not nil or named per-group fields otherwise.
     In another words, Equal() compares messages as if
     they were encoded